+++ Plugins/iOS/Editor/CommandRunner.cs
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2018, Oath Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using System.Diagnostics;
+using System;
+using System.IO;
+
+namespace FlurrySDK.Editor{
+    internal class ShellCommand {
+        static string AppendingHome(string path) {
+            var homePath = System.Environment.GetEnvironmentVariable("HOME");
+            return Path.Combine(homePath, path);
+        }
+
+        internal static void AddPossibleRubySearchPaths() {
+            AddSearchPath(AppendingHome(".rbenv/shims"));
+            AddSearchPath(AppendingHome(".rvm/scripts/rvm"));
+            AddSearchPath("/usr/local/bin/");
+        }
+
+        internal static void AddSearchPath(string path) {
+            var name = "PATH";
+            var currentPath = System.Environment.GetEnvironmentVariable(name);
+            var newPath = path + ":" + currentPath;
+            var target = EnvironmentVariableTarget.Process;
+            System.Environment.SetEnvironmentVariable(name, newPath, target);
+        }
+
+        internal static void Run(string command, string args) {
+            ProcessStartInfo psi = new ProcessStartInfo(); 
+            psi.FileName = command;
+            psi.UseShellExecute = false;
+            if (args != null) {
+                psi.Arguments = args;
+            }
+            Process p = Process.Start(psi); 
+            p.Close();
+        }
+    }
+}